migrated snackbar click callback test, fixed snapshot, removed duplicate enzyme test file
